+++ NEWS.md
@@ -1,6 +1,14 @@
 # cognitivemodels 0.0.10
 
+## New features
+
 * Added an option in `bayes()` to ignore the sum constraints of the prior parameter
+
+## Bugfixes
+
+* Fixed bugs in ROI optimization
+
+* Declared `solvers()` as depreciated with a warning, from now on use `cm_solvers()` to show the optimization solvers.
 
 # cognitivemodels 0.0.9